根据a0_lib修改03,04,05,并完成06

+++ assignment_0/src/rm_a0/a0_03_range_sum.hpp
@@ -1,6 +1,6 @@
 #ifndef A0_03_RANGE_SUM_HPP
 #define A0_03_RANGE_SUM_HPP
 namespace rm_a0{
-    int RangeSum(int l,int r);
+    long long RangeSum(long long l,long long r);
 }
 #endif
